Verifies creator, logged in user, and post exists to delete post

+++ handlers/deletepost.py
@@ -6,6 +6,11 @@ import time
 class DeletePost(PostPage):
     def post(self, post_id):
         post = Post.by_id(post_id)
-        post.delete()
-        time.sleep(0.1)
-        self.redirect("/")
+        creator = self.user.name == post.created_by or None
+
+        if self.user and post and creator:
+            post.delete()
+            time.sleep(0.1)
+            self.redirect("/")
+        else:
+            self.redirect("/login")
